Rework Glyph inheritance hierarchy to remove confusion and improve code-reuse.
The Glyph classes were becoming hard to understand because some classes
inherited the interface from more than one parent, while others used it
only as a marker interface. This change brings common behaviour up to Glyph,
which is now an abstract class. Code that might be specific to an Android
client is injected via the Grid interface. The old Grid class has become
DisplayGrid.

Another positive result is that the unit tests now test the production code
instead of a bunch of pointless test doubles.