Conversion to Android library project
I have done the basics of converting to an Android library project. I have
changed project.properties to make it a library project, and removed the
main activity and its layout. These UI pieces have been moved into another
Android project to prove the custom view can be used from the library.

I still need to remove more code that is not used, including an Activity,
and a lot of layout resources.