Build dot Drawables in code
The Drawables that represent the dim and lit dots are now built in code instead of being statically held in XML resources, allowing them to be configured by a
client.

This is not complete because at the moment the display is not rendering
properly. Dots set up correctly on the initial draw but behave very strangely
once they have been updated.