Completed color config via XML by fixing bugs and cleaning up code
I fixed the problem that the background and unchanged parts of digits were
drawing with default colors by moving the call to setFormat() to the end
of initialise(). This could still benefit from refactoring, because
setFormat() does much more than it should.
I also removed a lot of redundant logging code, and put a single piece of
logging into UpdateDigitsFromClockTask to help me to fix the next problem -
this AsyncTask queues itself to be re-run continuously, beyond the life of
the Activity, and fresh visits to the Activity just start up more tasks in
parallel.