Making dot size and padding between dots configurable from XML
The attributes are in place, and I have renamed the row and column padding
attributes. Code reads them. There is work still to do because they are
being converted from float to int, and dropping down to zero for smaller sizes.
I need to either implement a smallest size override (cannot be smaller than
1) or get a better understanding of how to work with dimensions.