Fixed bug with fade-in animation
Fade-in animation didn't work because it changed the dot after the fade-in.
This is correct for fade-out, but on fade-in this resulted in the dim dot
being faded from 0 to full alpha, showing through to the dim dot behind.
You could not see anything happen. Then the dot appeared.

The fix makes the dot change before the animation for a fade-in, and
changes the animation listener to only work for fade-outs.