Finished first cut of extended changeDot overload
Made ColorChangeAnimationListener configurable for next dot color and
whether the appear animation should run next. Finished writing first
cut of changeDot to use it. The code now feels like it is doing
all the right things, but is completely untested. Also feels like
it could be refactored to be clearer and not repeat itself.