Color changes now complete on the last dot update
Added code to changeDot() so that when the last dot - the one in the
bottom right corner - has been updated, the current colors are set
to the same value as the next colors.

This change has fixed the apparent problems with color cycling. The only
missing piece now is that spaces and colons don't change color.