Make colon and space able to redraw themselves for transitions
The space and colon now draw themselves with the same code as digits,
through the transition mechanism. This lets colons fade in with the
same speed as digits when the display is initially shown, and means
they can follow the identical dot rules to digits for transitioning
from one color to the next.