Lock color change at the start of setValue
I created a private field to indicate when a color change is active. It is set
to true if there is a pending color change at the start of setValue, and it
is cleared when changeDot draws the bottom right dot. changeDot has been
changed to use this field instead of making the next vs current comparison
for every dot.