Optimise setValue based on colorChangeInProgress flag
setValue now uses the new field to ensure it only redraws padding, spaces,
colons and digits that haven't changed if there is a color change currently
active.

This still hasn't fixed the problem of bits of the currently changing digit
from appearing too quickly. There must be a code reason for this, and it also
appears that this code has now hit a performance limit for using this method
to perform smooth animations on a color change.