Prove that the model can be used for drawing
I added a method to ModelGrid that doesn't belong to the Grid interface
to allow to get the state of any specific dot. Then I used the Grid
interface methods to set up a ModelGrid within SurfaceGridRendererThread
(setFormat() and setValue()) to set up a model with a the value
"93:34:07". This renders just like the Drawable version.