Moving towards dot color transitions
I have written code to calculate the per-frame fading/lightening colors, and
I converted the model to know about dots that are in transition.

The code hasn't been debugged, and very likely has errors. At the moment it
causes the display to flash betwen on and off.

This work is incomplete because it doesn't have code to update all the
transitioning dots to finished (lit or dim) dots during a setValue().