Progress with dot transitions
The code to get per-frame (time offset) dimming and lightening colors is
complete and in use, and the grid model is updating the transitioning dots
to dim or lit dots each time setValue() is called.

There is a problem somewhere in this code. The transitioning dots often
flip between fully on and off while others around them fade. Perhaps this
is a thread synchronization problem?