Fix flipping transitions
I fixed the transitions that were flipping between on and off by making
the call to clear the transition state happen immediately that the
transition has finished rather than the next time setValue is called. This
stops the transition being cleared part way into the next frame.

I also changed the timing of the call to setValue. The AsyncTask was
consistently running just before the second finished, rather than at
the start of the second. Most of the time the transition was considered
finished before it had even begun, so transition state was cleared immediately.