Extracted drawing code to pluggable strategy
I created an interface for objects that draw on a Canvas, extracted the
update-only and full drawing code to implementations of that interface,
and converted SurfaceGridRendererThread to use the interface.

This made it easy to perform a full draw once and do update-only draws
in the run() method. However, the update-only strategy seems to cause
the display to flicker on the dim dots. So the code checked in only ever
does full draws.