Model update moved into SurfaceGridRendererThread
The run() method main loop now checks for updates to the current value by
querying a configuration object of type UpdateProvider. It is also given
a timestamp for the earliest it should ask again, to avoid making many
unnecessary requests.

With this change there is no need for the old AsyncTask based code.