Integrated paint updating into SurfaceGridRendererThread
SurfaceGridRendererThread now takes a PaintUpdateProvider in the constructor
and makes use of it in setUpColors() and updatePaints() to enable colors to
change according to configuration.

SurfaceGridView now passes in a CountdownPaintUpdateProvider to cycle between
green, orange and red.

PaintUpdateProvider and CountdownPaintUpdate were changed so that int arrays
rather than Paint arrays are returned, as this better suits existing code in
SurfaceGridRendererThread.