Renamed interface, added the simple case SingleColorUpdateProvider
I renamed the interface PaintUpdateProvider to ColorUpdateProvider to better
reflect what it provides. I renamed other classes to match. I renamed one
interface method to make the same clarification.

Also I implemented a simple ColorUpdateProvider which returns only a single
color and when asked when the next update will be needed replies with,
effectively, the end of time.