Renamed classes for consistency
I noticed that the naming of the rendering classes was inconsistent with
the new name of the SurfaceView grid - MatrixDisplay. I renamed
MatrixDisplayRenderer to MatrixDisplayRenderController in acknowledgement
that it doesn't do any rendering - that has been moved out to implementations
of DrawStrategy. I also renamed the 2 DrawStrategy implementations to match.