Changes supporting name change and project/build cleanliness
In preparation for releasing the code I realised that I needed a clearer
name for the project. I changed the name of this, the test project and
the example project (formerly NumericViewHost).

Along the way I had problems building the projects in Eclipse, which lead
me to realise that they didn't build properly from the commandline. I have
updated the projects and made them reference each other properly.

I also noticed that the experimental code was extracted by copying all the
classes that the experimental code relies on too. I changed those projects
so that they reference this one. This doesn't show up in this commit but
it is out there.

The code is now ready to be merged into a new parent repo that can hold
all 3 subprojects - the view, the tests and the example. Perhaps I'll put
the experimental code in the parent too.