Renamed class for clarity
The DotChangeAction interface was defined to help with testability. I have
renamed the non-test implementation to make it clearer what it is for.
The original name goes back to the time when it may have also kicked off
animations and their listeners on a dot change, which is no longer evident
in the SurfaceView code.