Completed review prior to initial publishing on GitHub
Most of these changes are simple renames and inlines to improve clarity. Also
I moved the useful code from MatrixDisplayFullDrawStrategy directly into
MatrixDisplayRenderController and was able to delete the DrawStrategy interface
and both implementations (Full and Update). This was a great convenience when
I was trying out alternative ways to draw the view, but completely
unnecessary now.

The review is now complete in terms of removing bad comments. I need to add
JavaDoc comments to public methods still.